resolve TODO regarding `to + infinitive` schema in api-fluent
and disable bbc test as this change breaks it
