+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/anyExpectations.kt
@@ -5,6 +5,7 @@ import ch.tutteli.atrium.logic.creating.typeutils.IterableLike
 import ch.tutteli.atrium.logic.*
 import ch.tutteli.atrium.logic.creating.transformers.SubjectChangerBuilder
 import ch.tutteli.atrium.logic.utils.iterableLikeToIterable
+import ch.tutteli.atrium.reporting.Reporter
 import ch.tutteli.kbox.glue
 import kotlin.reflect.KClass
 
@@ -195,4 +196,45 @@ fun <T> Expect<T>.notToEqualOneOf(expected: T, vararg otherValues: T): Expect<T>
 fun <T> Expect<T>.notToEqualOneIn(expected: IterableLike): Expect<T> =
     _logicAppend { isNotIn(iterableLikeToIterable(expected)) }
 
+/**
+ * Allows to state a reason for one or multiple assertions for the current subject.
+ *
+ * @param reason The explanation for the assertion(s) created by [assertionCreator].
+ * @param assertionCreator The group of assertions to make.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.AnyExpectationSamples.because
+ *
+ * @since 0.15.0
+ */
+fun <T> Expect<T>.because(reason: String, assertionCreator: Expect<T>.() -> Unit): Expect<T> =
+    _logicAppend { because(reason, assertionCreator) }
 
+/**
+ * Can be used to separate single assertions.
+ *
+ * For instance `expect(1).isLessThan(2).and.isGreaterThan(0)` creates
+ * two assertions (not one assertion with two sub-assertions) - the first asserts that 1 is less than 2 and the second
+ * asserts that 1 is greater than 0. If the first assertion fails, then the second assertion is not evaluated.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.AnyExpectationSamples.andFeature
+ */
+inline val <T> Expect<T>.and: Expect<T> get() = this
+
+/**
+ * Can be used to create a group of sub assertions when using the fluent API.
+ *
+ * For instance `expect(1).isLessThan(3).and { isEven(); isGreaterThan(1) }` creates
+ * two assertions where the second one consists of two sub-assertions. In case the first assertion holds, then the
+ * second one is evaluated as a whole. Meaning, even though 1 is not even, it still evaluates that 1 is greater than 1.
+ * Hence the reporting might (depending on the configured [Reporter]) contain both failing sub-assertions.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.AnyExpectationSamples.and
+ */
+infix fun <T> Expect<T>.and(assertionCreator: Expect<T>.() -> Unit): Expect<T> =
+    _logic.appendAsGroup(assertionCreator)
