+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/collectionAssertions.kt
@@ -42,28 +42,3 @@ fun <T : Collection<*>> Expect<T>.isNotEmpty(): Expect<T> =
 @Deprecated("Use toHaveSize; will be removed with 1.0.0 at the latest", ReplaceWith("this.toHaveSize<T>(expected)"))
 fun <T : Collection<*>> Expect<T>.hasSize(expected: Int): Expect<T> =
     size { toEqual(expected) }
-
-//TODO move to collectionExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Collection.size] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.CollectionExpectationSamples.sizeFeature
- */
-val <T : Collection<*>> Expect<T>.size: Expect<Int>
-    get() = _logic.size(::identity).transform()
-
-//TODO move to collectionExpectations.kt with 0.18.0
-/**
- * Expects that the property [Collection.size] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.CollectionExpectationSamples.size
- */
-fun <E, T : Collection<E>> Expect<T>.size(assertionCreator: Expect<Int>.() -> Unit): Expect<T> =
-    _logic.size(::identity).collectAndAppend(assertionCreator)
