+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/creating/charsequence/contains/impl/StaticNames.kt
@@ -7,8 +7,7 @@ import ch.tutteli.atrium.logic.creating.charsequence.contains.steps.AtLeastCheck
 import kotlin.reflect.KFunction3
 
 internal object StaticNames {
-    //TODO rename with 0.18.0 to notToContain
-    val containsNotValuesFun = run {
+    val notToContainValuesFun = run {
         val f: KFunction3<Expect<CharSequence>, Any, Array<out Any>, Expect<CharSequence>> =
             Expect<CharSequence>::notToContain
         f.name
