+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/creating/iterable/contains/impl/StaticNames.kt
@@ -8,8 +8,7 @@ import ch.tutteli.atrium.logic.creating.iterable.contains.steps.AtLeastCheckerSt
 import kotlin.reflect.KFunction3
 
 internal object StaticNames {
-    //TODO rename with 0.18.0 notToContain
-    val containsNotValuesFun = run {
+    val notToContainValuesFun = run {
         val f: KFunction3<Expect<Iterable<Double>>, Double, Array<out Double>, Expect<Iterable<Double>>> =
             Expect<Iterable<Double>>::notToContain
         f.name
