+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/iterableLikeToContainCheckers.kt
@@ -1,7 +1,3 @@
-//TODO rename file to iterableLikeToContain... in 0.18.0
-//TODO remove both annotations with 1.0.0
-@file:JvmName("IterableContainsCheckers")
-
 package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.api.fluent.en_GB.creating.iterable.contains.impl.StaticNames
@@ -10,7 +6,6 @@ import ch.tutteli.atrium.logic._logic
 import ch.tutteli.atrium.logic.creating.iterable.contains.IterableLikeContains
 import ch.tutteli.atrium.logic.creating.iterable.contains.searchbehaviours.InAnyOrderSearchBehaviour
 import ch.tutteli.atrium.logic.creating.iterable.contains.steps.*
-import kotlin.jvm.JvmName
 
 /**
  * Restricts a `contains` assertion by specifying that the number of occurrences of the entry which we are looking
@@ -28,7 +23,7 @@ import kotlin.jvm.JvmName
 fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> IterableLikeContains.EntryPointStep<E, T, S>.atLeast(
     times: Int
 ): AtLeastCheckerStep<E, T, S> =
-    _logic.atLeastCheckerStep(times, StaticNames.containsNotValuesFun) { "${StaticNames.atLeast}($it)" }
+    _logic.atLeastCheckerStep(times, StaticNames.notToContainValuesFun) { "${StaticNames.atLeast}($it)" }
 
 /**
  * Restricts a `contains at least` assertion by specifying that the number of occurrences of the entry which we
@@ -51,7 +46,7 @@ fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> AtLeastCheckerStep<E, T,
     times: Int
 ): ButAtMostCheckerStep<E, T, S> = _logic.butAtMostCheckerStep(
     times,
-    StaticNames.containsNotValuesFun,
+    StaticNames.notToContainValuesFun,
     { l, u -> "${StaticNames.atLeast}($l).${StaticNames.butAtMost}($u)" },
     { "${StaticNames.atLeast}($it)" },
     { "${StaticNames.butAtMost}($it)" },
@@ -76,7 +71,7 @@ fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> AtLeastCheckerStep<E, T,
 fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> IterableLikeContains.EntryPointStep<E, T, S>.exactly(
     times: Int
 ): ExactlyCheckerStep<E, T, S> =
-    _logic.exactlyCheckerStep(times, StaticNames.containsNotValuesFun) { "${StaticNames.exactly}($it)" }
+    _logic.exactlyCheckerStep(times, StaticNames.notToContainValuesFun) { "${StaticNames.exactly}($it)" }
 
 /**
  * Restricts a `contains` assertion by specifying that the number of occurrences of the entry which we
@@ -100,7 +95,7 @@ fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> IterableLikeContains.Ent
     times: Int
 ): AtMostCheckerStep<E, T, S> = _logic.atMostCheckerStep(
     times,
-    StaticNames.containsNotValuesFun,
+    StaticNames.notToContainValuesFun,
     { "${StaticNames.atMost}($it)" },
     { "${StaticNames.atLeast}($it)" },
     { "${StaticNames.exactly}($it)" }
@@ -122,4 +117,4 @@ fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> IterableLikeContains.Ent
 fun <E, T: IterableLike, S : InAnyOrderSearchBehaviour> IterableLikeContains.EntryPointStep<E, T, S>.notOrAtMost(
     times: Int
 ): NotOrAtMostCheckerStep<E, T, S> =
-    _logic.notOrAtMostCheckerStep(times, StaticNames.containsNotValuesFun) { "${StaticNames.notOrAtMost}($it)" }
+    _logic.notOrAtMostCheckerStep(times, StaticNames.notToContainValuesFun) { "${StaticNames.notOrAtMost}($it)" }
