+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/mapLikeToContainSearchBehaviours.kt
@@ -1,4 +1,3 @@
-//TODO rename file to mapLikeToContain... in 0.18.0
 package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.logic.creating.typeutils.MapLike
