+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/throwableExpectations.kt
@@ -8,6 +8,29 @@ import ch.tutteli.atrium.logic.creating.transformers.SubjectChangerBuilder
 import kotlin.reflect.KClass
 
 /**
+ * Expects that the property [Throwable.message] of the subject of `this` expectation is not null,
+ * creates an [Expect] for it and returns it.
+ *
+ * @return The newly created [Expect] for the property [Throwable.message] of the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableExpectationSamples.message
+ */
+val <T : Throwable> Expect<T>.message: Expect<String>
+    get() = feature(Throwable::message).notToEqualNull()
+
+/**
+ * Expects that the property [Throwable.message] of the subject of `this` expectation is not null and
+ * holds all assertions the given [assertionCreator] creates for it and
+ * returns an [Expect] for the current subject of `this` expectation.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableExpectationSamples.messageFeature
+ */
+fun <T : Throwable> Expect<T>.message(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
+    feature(Throwable::message) { notToEqualNull(assertionCreator) }
+
+/**
  * Expects that the property [Throwable.message] of the subject of `this` expectation is not null and contains
  * [expected]'s [toString] representation and the [toString] representation of the [otherExpected] (if given),
  * using a non-disjoint search.
@@ -29,3 +52,39 @@ fun <T : Throwable> Expect<T>.messageToContain(
     expected: CharSequenceOrNumberOrChar,
     vararg otherExpected: CharSequenceOrNumberOrChar
 ): Expect<T> = message { toContain(expected, *otherExpected) }
+
+/**
+ * Expects that the property [Throwable.cause] of the subject *is a* [TExpected] (the same type or a sub-type),
+ * creates an [Expect] of the [TExpected] type for it and returns it.
+ *
+ * @return The newly created [Expect] for the property [Throwable.cause] of the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableExpectationSamples.causeFeature
+ *
+ * @since 0.10.0
+ */
+inline fun <reified TExpected : Throwable> Expect<out Throwable>.cause(): Expect<TExpected> =
+    causeIsInstanceOf(TExpected::class).transform()
+
+@PublishedApi // in order that _logic does not become part of the API we have this extra function
+internal fun <TExpected : Throwable> Expect<out Throwable>.causeIsInstanceOf(
+    kClass: KClass<TExpected>
+): SubjectChangerBuilder.ExecutionStep<Throwable?, TExpected> = _logic.causeIsA(kClass)
+
+/**
+ *
+ * Expects that the property [Throwable.cause] of the subject *is a* [TExpected] (the same type or a sub-type) and
+ * holds all assertions the given [assertionCreator] creates for it.
+ *
+ * Notice, in contrast to other expectation functions which expect an [assertionCreator], this function returns not
+ * [Expect] of the initial type, which was some type `T `, but an [Expect] of the specified type [TExpected].
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableExpectationSamples.cause
+ *
+ * @since 0.10.0
+ */
+inline fun <reified TExpected : Throwable> Expect<out Throwable>.cause(
+    noinline assertionCreator: Expect<TExpected>.() -> Unit
+): Expect<TExpected> = causeIsInstanceOf(TExpected::class).transformAndAppend(assertionCreator)
